saas_automation/requirements.txt
# External dependencies for SaaS Automation module
# Install these packages for full functionality

# Core dependencies (required for basic functionality)
docker>=6.0.0
paramiko>=3.0.0
psycopg2-binary>=2.9.0
cryptography>=3.4.0

# Optional dependencies (for advanced features)
kubernetes>=26.0.0  # For Kubernetes support
redis>=4.0.0        # For caching and session management
celery>=5.0.0       # For background tasks
requests>=2.28.0    # For HTTP requests

# Development dependencies (for testing and development)
pytest>=7.0.0
pytest-odoo>=1.0.0 